refactor: vITE migration pt 1
fixed type file clashes with VITE
moved index.html
migrated index.html to correlate with VITE standards
adjusted some types to match
adjusted package.json
removed @civic (issue's with vite)
replaced MintNest/NestQuestSingleListing with empty export
adjusted .less file wrt to antd
added vite.config.ts & vite-env.d.ts
added vite dependencies to package.json
.lock updated
