+++ index.html
@@ -3,16 +3,16 @@
   <head>
     <meta charset="utf-8" />
 
-    <script src="%PUBLIC_URL%/datafeeds/udf/dist/bundle.js"></script>
+    <script src="/datafeeds/udf/dist/bundle.js"></script>
     <script defer data-domain="app.goosefx.io" src="https://plausible.io/js/script.js"></script>
 
-    <link href="%PUBLIC_URL%/img/assets/g-logo.svg" rel="icon" />
-    <link href="%PUBLIC_URL%/img/assets/g-logo.svg" rel="appletouch-icon" />
+    <link href="/img/assets/g-logo.svg" rel="icon" />
+    <link href="/img/assets/g-logo.svg" rel="appletouch-icon" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="#ffffff" media="(prefers-color-scheme: light)" />
     <meta name="theme-color" content="#131313" media="(prefers-color-scheme: dark)" />
 
-    <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
+    <link rel="manifest" href="/manifest.json" />
 
     <title>GooseFX - Swap or Trade Crypto, Perpetuals, NFTs, & Liquidity Pools</title>
     <meta name="title" content="GooseFX - Swap or Trade Crypto, Perpetuals, NFTs, & Liquidity Pools" />
@@ -52,8 +52,8 @@
       integrity="sha512-+4zCK9k+qNFUR5X+cKL9EIR+ZOhtIloNl9GIKS57V1MyNsYpYcUrUeQc9vNfzsWfV28IaLL3i96P9sdNyeRssA=="
       crossorigin="anonymous"
     />
-    <link rel="stylesheet" type="text/css" href="%PUBLIC_URL%/css/nivo-lightbox/nivo-lightbox.css" />
-    <link rel="stylesheet" type="text/css" href="%PUBLIC_URL%/css/nivo-lightbox/default.css" />
+    <link rel="stylesheet" type="text/css" href="/css/nivo-lightbox/nivo-lightbox.css" />
+    <link rel="stylesheet" type="text/css" href="/css/nivo-lightbox/default.css" />
     <link href="https://fonts.googleapis.com/css?family=Open+Sans:300,400,600,700" rel="stylesheet" />
     <link href="https://fonts.googleapis.com/css?family=Lato:400,700" rel="stylesheet" />
     <link href="https://fonts.googleapis.com/css?family=Raleway:300,400,500,600,700,800,900" rel="stylesheet" />
@@ -78,19 +78,19 @@
       fbq('init', '490472842829687')
       fbq('track', 'PageView')
     </script>
-    <noscript>
-      <img
-        height="1"
-        width="1"
-        src="https://www.facebook.com/tr?id=490472842829687&ev=PageView
+    <img
+      height="1"
+      width="1"
+      src="https://www.facebook.com/tr?id=490472842829687&ev=PageView
     &noscript=1"
-      />
-    </noscript>
+    />
+    <noscript> </noscript>
     <!-- End Facebook Pixel Code -->
   </head>
 
   <body>
     <div id="root"></div>
     <script type="text/javascript" src="/js/jquery.1.11.1.js"></script>
+    <script type="module" src="/src/index.jsx"></script>
   </body>
 </html>
