+++ package.json
@@ -6,6 +6,8 @@
     "start": "PORT=3000 craco start",
     "build": "craco build",
     "test": "craco test",
+    "dev": "vite",
+    "vite:build": "tsc && vite build",
     "eject": "react-scripts eject",
     "format:fix": "prettier --write \"src/*.+(js|jsx|ts|tsx|json|css|md)\"",
     "format:check": "prettier -c \"src/*.+(js|jsx|ts|tsx|json|css|md)\"",
@@ -19,7 +21,6 @@
   "dependencies": {
     "@ant-design/plots": "^1.2.5",
     "@bonfida/spl-name-service": "^1.0.4",
-    "@civic/solana-gateway-react": "^0.7.5",
     "@headlessui/react": "^1.7.15",
     "@jup-ag/api": "^6.0.6",
     "@jup-ag/react-hook": "^6.0.0-beta.2",
@@ -51,15 +52,9 @@
     "axios": "^0.24.0",
     "bn.js": "^5.2.0",
     "borsh": "^0.6.0",
-    "browserify-zlib": "^0.2.0",
     "bs58": "^4.0.1",
-    "buffer": "^6.0.3",
-    "buffer-layout": "^1.2.2",
     "chart.js": "^3.5.1",
     "coingecko-api": "^1.0.10",
-    "crypto-browserify": "^3.12.0",
-    "customize-cra": "^1.0.0",
-    "dayjs": "^1.11.10",
     "echarts": "^5.1.2",
     "emailjs-com": "^3.2.0",
     "eventemitter3": "^4.0.7",
@@ -68,15 +63,12 @@
     "goosefx-stake-rewards-sdk": "^2.0.1",
     "highcharts": "^10.2.1",
     "highcharts-react-official": "^3.1.0",
-    "https-browserify": "^1.0.0",
     "jsonwebtoken": "8.5.1",
     "lodash.chunk": "^4.2.0",
     "lodash.orderby": "^4.6.0",
     "lottie-react": "^2.2.1",
     "moment": "^2.29.2",
     "openbook-ts": "file:./openbook-package",
-    "os-browserify": "^0.3.0",
-    "path-browserify": "^1.0.1",
     "perps-wasm": "file:./perps-wasm",
     "querystring-es3": "^0.2.1",
     "react": "^17.0.2",
@@ -93,14 +85,9 @@
     "react-slick": "^0.28.1",
     "slick-carousel": "^1.8.1",
     "smooth-scroll": "^16.1.3",
-    "stream-http": "^3.2.0",
     "styled-components": "5.1.1",
     "tailwindcss": "^3.0.24",
-    "ts-pattern": "^4.0.6",
-    "twin.macro": "^2.8.2",
-    "type": "^2.6.0",
-    "url": "^0.11.1",
-    "workbox-webpack-plugin": "^6.2.4"
+    "twin.macro": "^2.8.2"
   },
   "babelMacros": {
     "twin": {
@@ -127,8 +114,6 @@
   ],
   "devDependencies": {
     "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
-    "@craco/craco": "^7.1.0",
-    "@swc/core": "^1.3.84",
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^12.0.0",
     "@testing-library/user-event": "^13.2.1",
@@ -148,13 +133,10 @@
     "@types/tape": "^4.13.2",
     "@typescript-eslint/eslint-plugin": "^5.35.1",
     "@typescript-eslint/parser": "^5.35.1",
+    "@vitejs/plugin-react-swc": "^3.4.1",
     "arweave-deploy": "^1.9.1",
     "babel-plugin-macros": "^3.1.0",
     "babel-plugin-styled-components": "^2.1.4",
-    "compression-webpack-plugin": "^10.0.0",
-    "craco-antd": "^2.0.0",
-    "craco-esbuild": "^0.4.5",
-    "craco-less": "^3.0.1",
     "css-loader": "^6.8.1",
     "eslint": "^7.1.1",
     "eslint-config-airbnb": "19.0.4",
@@ -169,17 +151,18 @@
     "eslint-plugin-unused-imports": "^2.0.0",
     "gh-pages": "^3.2.3",
     "husky": "^7.0.2",
+    "less": "^4.2.0",
     "prettier": "^2.7.1",
     "pretty-quick": "^3.1.1",
-    "react-scripts": "^5.0.1",
     "source-map-loader": "^4.0.1",
     "spok": "^1.4.3",
     "style-loader": "^3.3.3",
     "tape": "^5.3.2",
-    "terser-webpack-plugin": "^5.3.9",
-    "wasm-loader": "^1.3.0",
-    "webpack-bundle-analyzer": "^4.9.1",
-    "yarn-upgrade-all": "^0.5.4"
+    "vite": "^4.5.0",
+    "vite-plugin-babel-macros": "^1.0.6",
+    "vite-plugin-node-polyfills": "^0.15.0",
+    "vite-plugin-top-level-await": "^1.3.1",
+    "vite-plugin-wasm": "^3.2.2"
   },
   "resolutions": {
     "@solana/web3.js": "^1.75.0",
