+++ src/pages/NFTs/Profile/HeaderProfile.tsx
@@ -7,7 +7,7 @@ import { Image } from 'antd'
 import { SkeletonCommon } from '../Skeleton/SkeletonCommon'
 import { generateTinyURL } from '../../../api/tinyUrl'
 import { useNFTProfile, useDarkMode, useConnectionConfig, useAccounts } from '../../../context'
-import { ILocationState } from '../../../types/app_params.d'
+import { ILocationState } from '../../../types/app_params'
 import { Button } from '../../../components'
 import { PopupProfile } from './PopupProfile'
 import { Share } from '../Share'
@@ -25,7 +25,7 @@ import { MainButton, SuccessfulListingMsg, FloatingActionButton } from '../../..
 import { checkMobile, notify, truncateAddress } from '../../../utils'
 import { StyledHeaderProfile, SETTLE_BALANCE_MODAL, MARGIN_VERTICAL } from './HeaderProfile.styled'
 import BN from 'bn.js'
-import { IAppParams } from '../../../types/app_params.d'
+import { IAppParams } from '../../../types/app_params'
 import { FLOATING_ACTION_ICON } from '../../../styles'
 import tw from 'twin.macro'
 import 'styled-components/macro'
