+++ src/pages/NFTs/Profile/PopupProfile.tsx
@@ -4,7 +4,7 @@ import { Form, Upload, UploadProps, Button } from 'antd'
 import { STYLED_PROFILE_POPUP } from './PopupProfile.styled'
 import { useNFTProfile, useDarkMode } from '../../../context'
 import tw from 'twin.macro'
-import { INFTProfile } from '../../../types/nft_profile.d'
+import { INFTProfile } from '../../../types/nft_profile'
 import { completeNFTUserProfile, updateNFTUser } from '../../../api/NFTs'
 import { CenteredDiv } from '../../../styles'
 import styled from 'styled-components'
