+++ src/types/buffer-layout.d.ts
@@ -1,4 +1,4 @@
-declare module 'buffer-layout' {
+declare module 'src/types/buffer-layout' {
   const magic: any
   export = magic
 }
