+++ src/types/promise.ts
