+++ src/web3/nfts/index.ts
@@ -1,6 +1,6 @@
 export * from './metadata'
 export * from './mintNFT'
-export * from './types.d'
+export * from './types'
 export { getParsedNftAccountsByOwner } from './getParsedNftAccountsByOwner'
 export { getParsedNftAccountsByUpdateAuthority } from './getParsedNftAccountsByUpdateAuthority'
 export { getParsedAccountByMint } from './getParsedAccountByMint'
