+++ vite.config.ts
@@ -1,0 +1,22 @@
+import { defineConfig } from 'vite'
+import react from '@vitejs/plugin-react-swc'
+import macrosPlugin from 'vite-plugin-babel-macros'
+import { nodePolyfills } from 'vite-plugin-node-polyfills'
+import wasm from 'vite-plugin-wasm'
+import topLevelAwait from 'vite-plugin-top-level-await'
+// https://vitejs.dev/config/
+export default defineConfig({
+  base: '/',
+  plugins: [react(), macrosPlugin(), nodePolyfills(), wasm(), topLevelAwait()],
+  css: {
+    preprocessorOptions: {
+      less: {
+        javascriptEnabled: true
+      }
+    }
+  },
+  define: {
+    'process.env': {},
+    global: 'window'
+  }
+})
